requirements.txt
# Web Scraping
google-play-scraper==1.2.2

# Data Processing
pandas==2.1.4
numpy==1.26.2

# NLP and Sentiment Analysis
transformers==4.36.2
torch==2.1.2
vaderSentiment==3.3.2
textblob==0.17.1
spacy==3.7.2

# Machine Learning
scikit-learn==1.3.2

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0

